Issue #3985: Move all input files for EmptyLineSeparator check
