--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/whitespace/InputPrePreviousLineEmptiness.java
@@ -1,6 +1,0 @@
-
-package com.puppycrawl.tools.checkstyle.checks.whitespace;
-
-public class InputPrePreviousLineEmptiness {
-
-}
