+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/whitespace/emptylineseparator/InputEmptyLineSeparatorMultipleEmptyLines.java
@@ -18,7 +18,7 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 
-package com.puppycrawl.tools.checkstyle.checks.whitespace;
+package com.puppycrawl.tools.checkstyle.checks.whitespace.emptylineseparator;
 
 
 import java.util.*;
