+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/whitespace/emptylineseparator/InputEmptyLineSeparatorMultipleImportEmptyClass.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks.whitespace;import java.util.List;
+package com.puppycrawl.tools.checkstyle.checks.whitespace.emptylineseparator;import java.util.List;
 
 import java.util.Calendar;
 import java.util.Date;
