+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/whitespace/emptylineseparator/package-info/test1/package-info.java
@@ -1,5 +1,5 @@
 /*OK: for test allowing to place annotation before PACKAGE_DEF.*/
 @Deprecated
-package com.puppycrawl.tools.checkstyle.checks.whitespace.packageinfo.test1;
+package com.puppycrawl.tools.checkstyle.checks.whitespace.emptylineseparator.packageinfo.test1;
 
 import java.lang.Deprecated;
