+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/whitespace/emptylineseparator/package-info/test2/package-info.java
@@ -1,5 +1,5 @@
 /**OK: for test allowing to place javadoc before PACKAGE_DEF.*/
-package com.puppycrawl.tools.checkstyle.checks.whitespace.packageinfo.test2;
+package com.puppycrawl.tools.checkstyle.checks.whitespace.emptylineseparator.packageinfo.test2;
 
 //that import is for doing check of PACKAGE_DEF,
 //because EmptyLineSeparatorCheck doesn't check last token
